managed(???) a shape. kind of.
managed to get the overall hexagon shape. Might have to rethink the actual tile shape because the <BR> space can't be eliminated in a way I want. Tried [line-height] but due to the shape of the tiles intersecting, (the hexagon tips/edges intersecting), setting the value to 0 still does not remove the whitespace. Will investigate further w/ other CSS properties/tile shapes.