spacing, text
played around with spacing around various elements such as the nav, the main page itself and added some effects (hover, activated) to the menu buttons. Embedded an instagram post for the requirement of using an "external service. Drabbled some quick lines for the main and contact page, to (gradually) replace the lorem ipsum.